optimized and added comments

+++ Form_NSST/Form_Empty/FFT.h
@@ -1,3 +1,3 @@
 #pragma once
-void fft(int n, double* xRe, double* xIm, double* yRe, double* yIm);
-void ifft(int n, double* xRe, double* xIm, double* yRe, double* yIm);
+void fft(int n, float* xRe, float* xIm, float* yRe, float* yIm);
+void ifft(int n, float* xRe, float* xIm, float* yRe, float* yIm);
