+++ Form_NSST/Form_Empty/Process.h
@@ -1,10 +1,10 @@
 #pragma once
 #include <Windows.h>
 struct imagener {
-	double* Real;
-	double* Im;
+	float* Real;
+	float* Im;
 	int w;
 	int h;
 };
-void FFT2D(double* img, double* out_real, double* out_imag, int width, int height);
-void IFFT2D(double* Output_real, double* Output_imag, double* Input_real, double* Input_imag, int width, int height);
+void FFT2D(float* img, float* out_real, float* out_imag, int width, int height);
+void IFFT2D(float* Output_real, float* Output_imag, float* Input_real, float* Input_imag, int width, int height);
